+++ test/wallets/test_report_overview.js
@@ -21,7 +21,7 @@ const report = [
   },
   {},
   {
-    subtitle: '50 years ago',
+    subtitle: '51 years ago',
     title: 'Last Block:',
   },
   {